% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/puffer.R
\docType{data}
\name{puffer}
\alias{puffer}
\title{Dataset on pufferfish species and their fossil diversities.}
\format{One phylo object and one dataframe}
\description{
A dataset with a phylogenetic tree and fossil diversity data for pufferfishes
}
\details{
\itemize{
  \item pufferTree. Phylogenetic tree of pufferfish
  \item pufferRichness. Includes only extant taxa
}
}
\keyword{datasets}

