% Generated by roxygen2 (4.0.0): do not edit by hand
\name{runFossilMEDUSA}
\alias{runFossilMEDUSA}
\title{runFossilMEDUSA: MEDUSA including fossil diversity data}
\usage{
runFossilMEDUSA(phy, richness = NULL, fossil.richness = NULL,
  est.extinction = T, fossil.minimum = F, model.limit = 20,
  epsilon.value = NULL, mc = F, ...)
}
\arguments{
\item{phy}{An object of class 'phylo'}

\item{richness}{Richness data for extant species}

\item{fossil.richness}{Fossil richness data}

\item{est.extinction}{Estimate extinction rates or not?}

\item{fossil.minimum}{Treat fossils as minimum diversity estimates? If F, numbers are treated as true estimates}

\item{model.limit}{Maximum number of split models to consider}

\item{epsilon.value}{Input fixed epsilon value; if NULL, then epsilon is estimated}

\item{mc}{Use multicore or not}
}
\value{
List with elements
 $models, which contains:
    $par: i x 2 matrix (for birth-death; i x 1 for pure-birth); the jth row contains
      the speciation and (optional) extinction rate for the jth rate class
    $lnLik.part: vector of length i; the jth element is the partial log
      likelihood value due to the jth rate class
    $lnLik: = sum(lnLik.part); the overall log-likelihood of this model
    $split.at: the i+1 locations of splits.  The first element is the root node (i.e. background rate).
 $phy: the tree, which may have been manipulated (renamed, pruned) in the analysis
 $z: a matrix listing branch times and richnesses; for summarizing results
 $anc: a list of all ancestors; for summarizing results
 $model.summary: data.frame listing considered models, likelihoods, AIC, etc.

The species richness information is assumed to have columns "taxon" and "n.taxa";
"taxon" must match with a tip.label in the phylogeny "phy". May also include 'exemplar' column,
used for incompletely-sampled clades which requiring pruning.
}
\description{
This function runs MEDUSA but includes data on fossil diversities through time. Requires
phylogeny 'phy' and (optional) 'richness' (containing extant and potentially fossil
counts); 'fossil.richness' optionally passed in separately. If no richness information
is provided then it is assumed tips represent single species with comprehensive
sampling. Returns a list of models with 0, 1, 2, ..., 'model.limit' partitions.
'model.limit' can be supplied by user, but is overruled if large enough (relative
to tree size) that AIC correction factor becomes undefined (different for pure birth vs.
birth-death models). 'phy' is assumed to be ultrametric and rescaled (i.e. calibrated) to time.
}

