% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/whale.R
\docType{data}
\name{whale}
\alias{whale}
\title{Dataset on whales and their fossil diversities.}
\format{Two phylo objects and three dataframes}
\description{
A dataset with a phylogenetic tree and fossil diversity data for whales.
}
\details{
\itemize{
  \item whale.tree. Phylogenetic tree of whales, 84 tips
  \item whale.pruned.tree. Phylogenetic tree of whales pruned to match richness files, 43 tips
  \item whale.richness. Includes both fossils and extant taxa
  \item whale.ext.richness. Richness of extant clades
  \item whale.fossil.richness. Richness of fossil clades
}
}
\keyword{datasets}

