% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{fossilMEDUSA}
\alias{fossilMEDUSA}
\alias{fossilMEDUSA-package}
\title{fossilMEDUSA: a package that adds fossils to MEDUSA}
\description{
The fossilMEDUSA package is a set of functions that can be used to run fossilMEDUSA
}
\author{
Luke J. Harmon, Joseph W. Brown, Rich FitzJohn, Mike Alfaro
}

