% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{mustelid}
\alias{mustelid}
\title{Dataset on mustelids and their fossil diversities.}
\format{One phylo objects and two dataframes}
\description{
A dataset with a phylogenetic tree and fossil diversity data for mustelids.
}
\details{
\itemize{
  \item mustelid.tree. Phylogenetic tree of whales, 46 tips
  \item mustelid.richness. Includes both fossils and extant taxa
  \item mustelid.ext.richness. Richness of extant clades
}
}
\keyword{datasets}

